+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to show total time after all exercises
+- Configuration to show correct answers after each exercise or after all
 
 ### Changed